move Value and ValueImpl
